feat: rename divider to line
