+++ example/primitives.elos
@@ -1,3 +1,74 @@
+def preview "Welkom op onze nieuwsbrief"
+def width 750
+def hgap 24
+def vgap 24
+
+style txt {
+    size 20
+    color "#333333"
+}
+
+style .hh {
+    size 20
+}
+
+style btn {
+    bgcolor "#FF00FF"
+    padding 20
+    size 18
+}
+
 body {
-    img "https://dummyimage.com/600x400/000/fff" -> "something"
+    space
+    img "https://dummyimage.com/600x200/000/fff" -> "https://www.google.be"
+    group {
+        txt "Hallo Sjouwke!"
+        cols
+            col {
+                img "https://dummyimage.com/600x400/000/fff"
+                line
+            }
+            col {
+                img "https://dummyimage.com/600x400/000/fff"
+                line
+            }
+            col {
+                img "https://dummyimage.com/600x400/000/fff"
+                line
+            }
+    }
+    group.cta {
+        cols
+            col {
+                txt "Een titel"
+                btn "Ontdek ons aanbod" -> "https://www.reinvanoyen.be"
+            }
+            col {
+                txt "Een titel"
+                 btn "Ontdek ons beter aanbod" -> "https://www.sjouwke.be"
+            }
+    }
+    group {
+        cols
+            col {
+                txt "Hallo dit is leuk"
+                img "https://dummyimage.com/600x400/000/fff"
+            }
+            col {
+                txt "Hallo dit is leuk"
+                img "https://dummyimage.com/600x400/000/fff"
+            }
+    }
+    group.footer {
+        cols
+            col {
+                txt.hh "Hallo dit is leuk"
+                img "https://dummyimage.com/600x400/000/fff"
+            }
+            col {
+                txt.hh "Hallo dit is leuk"
+                img "https://dummyimage.com/600x400/000/fff"
+            }
+    }
+    space
 }
